requirements.txt
google-genai
python-dotenv
pydantic
fastapi
uvicorn
psycopg2-binary
